test(apt): expect currently failing test to fail (#286)
One of the `apt` functional tests is currently failing, presumably due
to changes in the external repository we were using for this test. This
is tracked in #282. As a temporary quality of life improvement for
everyone who merges in this repository (as this test is run on every
merge to `main`) this PR marks this test with `pytest.mark.xfail`.