ci: add maintenance branch to workflows triggers to minimise diff (#288)
This PR adds the `*-maintenance` branches to the list of branches we
trigger workflows from on `push`. This is in preparation for updating
the CI for the `20.04-maintenance` branch, which is necessary if we want
to publish from that branch, since we've switched from using Github
releases to releasing via the `publish.yaml` workflow (#287).