ci: pull in changes made to CI in main (#287)
The repo CI has changed a lot since the `20.04-maintenance` branch was
created. The primary of this PR is to update the release CI so that we
can release from this branch, as it was created when we used Github
releases. To do this while minimising divergences from `main`, this PR
essentially copies the updated CI files from `main`.

The following files are added from `main` exactly as-is:
- `.github` directory, minus changes to the
`get-supported-python-versions.py` script (so we still test on Python
3.8)
- `.scripts` directory
- top-level CI files (`justfile`, `interface.just`,
`test-requirements.txt`)
- template and examples, since CI checks they work (`.template`,
`.example`, `interfaces/.example`)

Additionally, the following changes are made:
- keeping the `requires-python` in the top-level `pyproject.toml` at
`>=3.8`
- generating `interfaces/index.json` via the automation
- cherry-picking #286

The zizmor errors are present on main currently, and tracked in #240.