ci: pull in changes to docs CI from main so linkcheck won't auto-fail (#290)
This PR pulls in changes to the docs CI from `main` as well, which was
neglected in #287. We run `linkcheck` with slightly different arguments
on `push` than we do on `pull_request`, so this wasn't caught in the PR.
This PR updates the `docs.just` file to match the copy in `main`. The
relevant change is the extra variadic argument for the `linkcheck`
recipe.