feat: release charmlibs.snap 0.8.0 with Python 3.8 support (#291)
This PR adds a release of `charmlibs.snap` with Python 3.8 support. The
version is 0.8.0, as we want a sub 1.0 version to convey the lack of
long term support, and so that it doesn't show up as the latest release
on PyPI. The differences from the latest (1.0.1) version are some
`typing` changes and the required Python version.

We can't run the functional tests on Ubuntu 20.04 easily, as Github has
discontinued that runner, but the unit tests pass as expected when run
with Python 3.8.

Will address #283.