ci: fix old version check for adding a package to an existing interface (#295)
In `main`, the `ls.py` script run with the `--only-if-version-changed`
flag tries to get the old artefact (interface or package) version as
long as the directory (e.g. `interfaces/foo`) existed in the old commit.
This fails in CI when we the new commit adds a Python package to an
existing interface definition, because the directory exists in the old
commit, but the expected `pyproject.toml` files doesn't.

(There would be a symmetrical failure for adding an interface definition
to an existing package, but we don't run anything conditional on that in
CI.)

This PR fixes this bug by skipping based on the relevant
`FileNotFoundError` instead of testing for the directory's presence.