ci: use TarFile.extractall in the recommended way (#297)
This PR updates `.scripts/ls.py` to use `TarFile.extractall` in the
[recommended
way](https://docs.python.org/3.12/library/tarfile.html#tarfile.TarFile.extractall)
-- setting `filter='data'`. This silences a deprecation warning for
using `extractall` without an explicit `filter` argument, and removes
the need to `noqa`
[S202](https://docs.astral.sh/ruff/rules/tarfile-unsafe-members/)
(though we trust the extracted data anyway since it's a snapshot of our
own repo). This raises the minimum Python version needed to run the
script, as the `filter` argument was added in Python 3.12, but I believe
we were running with this version in CI anyway (which is why we were
getting the deprecation warnings).